requirements.txt
torch>=1.7.0
torchvision>=0.8.0
Pillow